# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/risks": {
      "get": {
        "tags": [
          "risks"
        ],
        "summary": "Get Risks",
        "description": "List the risk forecast queue within the caller's scope.\n\nArgs:\n    facility_id: Optional exact facility filter.\n    risk_type: Optional exact RISK_TYPES filter.\n    risk_level: Optional exact RISK_LEVELS filter.\n    decision_status: Optional exact decision_status filter.\n    from_: Optional ISO 8601 inclusive lower bound on as_of.\n    to: Optional ISO 8601 inclusive upper bound on as_of.\n    cursor: Opaque pagination cursor.\n    limit: Page size, 1-200.\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    A RiskListEnvelope sorted priority_score desc, prediction_window_start asc.",
        "operationId": "get_risks_api_v1_risks_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "risk_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Type"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "decision_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision Status"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{risk_id}": {
      "get": {
        "tags": [
          "risks"
        ],
        "summary": "Get Risk",
        "description": "Fetch one risk's full forecast card, enforcing the caller's scope.\n\nArgs:\n    risk_id: The requested risk id.\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    The risk's full detail card.\n\nRaises:\n    ApiError: 404 if the risk does not exist; 403 if it exists but is\n        outside scope.",
        "operationId": "get_risk_api_v1_risks__risk_id__get",
        "parameters": [
          {
            "name": "risk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Risk Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{risk_id}/acknowledge": {
      "post": {
        "tags": [
          "risks"
        ],
        "summary": "Acknowledge Risk",
        "description": "Acknowledge a risk forecast, with optimistic concurrency.\n\nArgs:\n    risk_id: The risk to acknowledge.\n    body: The expected_version the caller last saw.\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    The updated risk card.\n\nRaises:\n    ApiError: 404/403 per scope (see get_risk_detail); 409 if\n        expected_version is stale.",
        "operationId": "acknowledge_risk_api_v1_risks__risk_id__acknowledge_post",
        "parameters": [
          {
            "name": "risk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Risk Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{risk_id}/reject": {
      "post": {
        "tags": [
          "risks"
        ],
        "summary": "Reject Risk",
        "description": "Reject a risk forecast; requires a reason_code, and a comment for \"other\".\n\nArgs:\n    risk_id: The risk to reject.\n    body: expected_version, reason_code (from REJECT_REASONS), and an\n        optional comment (required when reason_code=\"other\").\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    The updated risk card.\n\nRaises:\n    ApiError: 422 for a missing/unknown reason_code, or a missing\n        comment when the reason requires one; 404/403 per scope; 409\n        if expected_version is stale.",
        "operationId": "reject_risk_api_v1_risks__risk_id__reject_post",
        "parameters": [
          {
            "name": "risk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Risk Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risks/{risk_id}/defer": {
      "post": {
        "tags": [
          "risks"
        ],
        "summary": "Defer Risk",
        "description": "Defer a risk forecast, with optimistic concurrency.\n\nArgs:\n    risk_id: The risk to defer.\n    body: The expected_version the caller last saw.\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    The updated risk card.\n\nRaises:\n    ApiError: 404/403 per scope; 409 if expected_version is stale.",
        "operationId": "defer_risk_api_v1_risks__risk_id__defer_post",
        "parameters": [
          {
            "name": "risk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Risk Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcknowledgeRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "AcknowledgeRequest",
        "description": "Body for POST /api/v1/risks/{risk_id}/acknowledge."
      },
...
      "DeferRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "DeferRequest",
        "description": "Body for POST /api/v1/risks/{risk_id}/defer."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionWindow": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "PredictionWindow",
        "description": "The forecast's predicted incident window (as_of + lead_min_hours .. + horizon_hours)."
      },
...
      "RejectRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "RejectRequest",
        "description": "Body for POST /api/v1/risks/{risk_id}/reject.\n\nreason_code is Optional at the schema level -- its presence is a\ndomain rule (contract: \"reject requires reason_code\"), enforced as a\n422 DOMAIN_VALIDATION_ERROR by the endpoint, not a 400 structural\nvalidation error."
      },
...
      "RiskListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RiskOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/RiskListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "RiskListEnvelope",
        "description": "The standard {data, meta} envelope for GET /api/v1/risks."
      },
      "RiskListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "RiskListMeta",
        "description": "Cursor-pagination metadata for the risk list envelope."
      },
      "RiskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "forecast_id": {
            "type": "string",
            "title": "Forecast Id"
          },
          "risk_type": {
            "type": "string",
            "title": "Risk Type"
          },
          "target": {
            "$ref": "#/components/schemas/RiskTarget"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "lead_min_hours": {
            "type": "number",
            "title": "Lead Min Hours"
          },
          "horizon_hours": {
            "type": "number",
            "title": "Horizon Hours"
          },
          "prediction_window": {
            "$ref": "#/components/schemas/PredictionWindow"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "decision_status": {
            "type": "string",
            "title": "Decision Status"
          },
          "sla_due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sla Due At"
          },
          "data_health": {
            "type": "string",
            "title": "Data Health"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "top_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Factors"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "forecast_id",
          "risk_type",
          "target",
          "as_of",
          "lead_min_hours",
          "horizon_hours",
          "prediction_window",
          "probability",
          "threshold",
          "risk_level",
          "priority_score",
          "decision_status",
          "sla_due_at",
          "data_health",
          "model",
          "top_factors",
          "recommendation",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "RiskOut",
        "description": "The full risk forecast card."
      },
      "RiskTarget": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "facility_id"
        ],
        "title": "RiskTarget",
        "description": "What this forecast is about -- always a sensor in this ticket's scope."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}